requirements.txt
spidev==3.8
gpiozero==2.0.1
lgpio==0.2.2.0
pillow==12.0.0
pathlib==1.0.1
playwright==1.56.0